requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
connectorx==0.3.1
Django==4.2
numpy==1.24.2
pandas==2.0.0
polars==0.16.18
pyarrow==11.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
whitenoise==6.4.0
XlsxWriter==3.0.9
